move container networking to own network.go file
