+++ pkg/util/docker/docker.go
@@ -8,17 +8,12 @@
 package docker
 
 import (
-	"bufio"
 	"bytes"
 	"context"
-	"encoding/binary"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"net"
 	"os"
-	"path/filepath"
-	"strconv"
 	"strings"
 	"time"
 
@@ -26,7 +21,6 @@ import (
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/client"
 
-	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/util/cache"
 )
 
@@ -66,36 +60,6 @@ const (
 // HostnameProvider docker implementation for the hostname provider
 func HostnameProvider(hostName string) (string, error) {
 	return GetHostname()
-}
-
-// DefaultGateway returns the default Docker gateway.
-func DefaultGateway() (net.IP, error) {
-	procRoot := config.Datadog.GetString("proc_root")
-	netRouteFile := filepath.Join(procRoot, "net", "route")
-	f, err := os.Open(netRouteFile)
-	if os.IsNotExist(err) || os.IsPermission(err) {
-		log.Errorf("unable to open %s: %s", netRouteFile, err)
-		return nil, nil
-	} else if err != nil {
-		// Unknown error types will bubble up for handling.
-		return nil, err
-	}
-	defer f.Close()
-
-	ip := make(net.IP, 4)
-	scanner := bufio.NewScanner(f)
-	for scanner.Scan() {
-		fields := strings.Fields(scanner.Text())
-		if len(fields) >= 3 && fields[1] == "00000000" {
-			ipInt, err := strconv.ParseInt(fields[2], 16, 32)
-			if err != nil {
-				return nil, fmt.Errorf("unable to parse ip %s, from %s: %s", fields[2], netRouteFile, err)
-			}
-			binary.LittleEndian.PutUint32(ip, uint32(ipInt))
-			break
-		}
-	}
-	return ip, nil
 }
 
 // ContainerIDToEntityName returns a prefixed entity name from a container ID
