+++ pkg/util/docker/network.go
@@ -1,0 +1,142 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2017 Datadog, Inc.
+
+// +build docker
+
+package docker
+
+import (
+	"bufio"
+	"encoding/binary"
+	"fmt"
+	"net"
+	"os"
+	"path/filepath"
+	"sort"
+	"strconv"
+	"strings"
+
+	log "github.com/cihub/seelog"
+	"github.com/docker/docker/api/types"
+
+	"github.com/DataDog/datadog-agent/pkg/config"
+)
+
+type dockerNetwork struct {
+	iface      string
+	dockerName string
+}
+
+type dockerNetworks []dockerNetwork
+
+func (a dockerNetworks) Len() int           { return len(a) }
+func (a dockerNetworks) Swap(i, j int)      { a[i], a[j] = a[j], a[i] }
+func (a dockerNetworks) Less(i, j int) bool { return a[i].dockerName < a[j].dockerName }
+
+var hostNetwork = dockerNetwork{"eth0", "bridge"}
+
+func findDockerNetworks(containerID string, pid int, netSettings *types.SummaryNetworkSettings) []dockerNetwork {
+	// Verify that we aren't using an older version of Docker that does
+	// not provider the network settings in container inspect.
+	if netSettings == nil || netSettings.Networks == nil {
+		log.Debugf("No network settings available from docker, defaulting to host network")
+		return []dockerNetwork{hostNetwork}
+	}
+
+	var err error
+	dockerGateways := make(map[string]int64)
+	for netName, netConf := range netSettings.Networks {
+		gw := netConf.Gateway
+		if netName == "host" || gw == "" {
+			log.Debugf("Empty network gateway, container %s is in network host mode, its network metrics are for the whole host", containerID)
+			return []dockerNetwork{hostNetwork}
+		}
+
+		// Check if this is a CIDR or just an IP
+		var ip net.IP
+		if strings.Contains(gw, "/") {
+			ip, _, err = net.ParseCIDR(gw)
+			if err != nil {
+				log.Warnf("Invalid gateway %s for container id %s: %s, skipping", gw, containerID, err)
+				continue
+			}
+		} else {
+			ip = net.ParseIP(gw)
+			if ip == nil {
+				log.Warnf("Invalid gateway %s for container id %s: %s, skipping", gw, containerID, err)
+				continue
+			}
+		}
+
+		// Convert IP to int64 for comparison to network routes.
+		dockerGateways[netName] = int64(binary.BigEndian.Uint32(ip.To4()))
+	}
+
+	// Read contents of file. Handle missing or unreadable file in case container was stopped.
+	procNetFile := hostProc(strconv.Itoa(int(pid)), "net", "route")
+	if !pathExists(procNetFile) {
+		log.Debugf("Missing %s for container %s", procNetFile, containerID)
+		return nil
+	}
+	lines, err := readLines(procNetFile)
+	if err != nil {
+		log.Debugf("Unable to read %s for container %s", procNetFile, containerID)
+		return nil
+	}
+	if len(lines) < 1 {
+		log.Errorf("empty network file, unable to get docker networks: %s", procNetFile)
+		return nil
+	}
+
+	networks := make([]dockerNetwork, 0)
+	for _, line := range lines[1:] {
+		fields := strings.Fields(line)
+		if len(fields) < 8 {
+			continue
+		}
+		if fields[0] == "00000000" {
+			continue
+		}
+		dest, _ := strconv.ParseInt(fields[1], 16, 32)
+		mask, _ := strconv.ParseInt(fields[7], 16, 32)
+		for net, gw := range dockerGateways {
+			if gw&mask == dest {
+				networks = append(networks, dockerNetwork{fields[0], net})
+			}
+		}
+	}
+	sort.Sort(dockerNetworks(networks))
+	return networks
+}
+
+// DefaultGateway returns the default Docker gateway.
+func DefaultGateway() (net.IP, error) {
+	procRoot := config.Datadog.GetString("proc_root")
+	netRouteFile := filepath.Join(procRoot, "net", "route")
+	f, err := os.Open(netRouteFile)
+	if os.IsNotExist(err) || os.IsPermission(err) {
+		log.Errorf("unable to open %s: %s", netRouteFile, err)
+		return nil, nil
+	} else if err != nil {
+		// Unknown error types will bubble up for handling.
+		return nil, err
+	}
+	defer f.Close()
+
+	ip := make(net.IP, 4)
+	scanner := bufio.NewScanner(f)
+	for scanner.Scan() {
+		fields := strings.Fields(scanner.Text())
+		if len(fields) >= 3 && fields[1] == "00000000" {
+			ipInt, err := strconv.ParseInt(fields[2], 16, 32)
+			if err != nil {
+				return nil, fmt.Errorf("unable to parse ip %s, from %s: %s", fields[2], netRouteFile, err)
+			}
+			binary.LittleEndian.PutUint32(ip, uint32(ipInt))
+			break
+		}
+	}
+	return ip, nil
+}
